pyproject.toml
[tool.poetry]
name = "searchright-technical-assignment"
version = "0.1.0"
description = ""
authors = ["Nuung <qlgks1@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
psycopg2-binary = "^2.9.10"
openai = "^1.75.0"
django = "^5.2.4"
djangorestframework = "^3.16.0"
psycopg2 = "^2.9.10"
pgvector = "^0.4.1"
tiktoken = "^0.9.0"
pytest = "^8.4.1"
pydantic-settings = "^2.10.1"
pyyaml = "^6.0.2"
dynaconf = "^3.2.11"
jinja2 = "^3.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
